AITRIOS/requirements.txt
requests
opencv-python
pillow
flatbuffers
GPIO
rpi.gpio
spidev
st7789